requirements.txt
flask
torch
open_clip_torch
numpy
scikit-learn
pillow
pandas